Add AC_CASE for aarch64 host_cpu
The 64-bit version of raspi os is not able to compile openocd-spi with building support for SPI on BCM2835 like the 32-bit arm version can.
In order to make it work, a new AC_CASE with the contents of the arm case needs to be added. The case name needs to be changed to [aarch64*] and after that openocd-spi can be built with support for SPI on BCM2835.