wip: Update BCM2835SPI support
I have been unable to get the rebased driver to work.
i assume this is related to the buffering changes to the bitbang driver since the fork.

See:
e971c7f42 (Add SWD protocol support to bitbang, 2014-12-11)

some basic communication is happening, as openocd is correctly receiving the initial dap id from the target.
The next thing to do would be to explore wether some of the "unused" functions
in the spi driver are actually used